pyproject.toml
[project]
name = "Easy-Parking"
version = "0.3.0"
description = "Parking equipment shop"
requires-python = ">=3.12"
dependencies = [
    "django==5.2.1",

    "django-phonenumber-field==8.1.0",
    "django-cotton==2.1.2",
    "django-simple-history==3.8.0",
    "django-structlog==9.1.1",
    "django-storages==1.14.6",
    "django-constance==4.3.2",
    "django-appconf==1.1.0",
    "django-cacheops==7.2",
    "django-htmx>=1.23.0",

    "celery[django,redis,librabbitmq]==5.5.2",
    "sqlalchemy==2.0.20",
    "flower==2.0.1",
    "redis==5.3.0",
    "django-celery-beat==2.8.0",

    "phonenumbers>=9.0.5",
    "pillow==11.2.1",
    "psycopg[binary,pool]>=3.2.8",
    "boto3==1.38.13",
    "dotenv==0.9.9",
    #    "gunicorn==23.0.0",
    "granian>=2.2.6",
]

[dependency-groups]
dev = [
       "django-debug-toolbar==5.2.0",
]

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    # Found commented-out code
    "ERA001",
]

[tool.ruff]
line-length = 119
